requirements.txt
# AI-Powered ATS-Friendly Resume Builder - Backend Dependencies
# Phase 1: Foundation and Compliance Backend (Enhanced)

# Core Framework
fastapi==0.115.5
pydantic==2.10.3
pydantic-settings==2.6.1

# ASGI Server
uvicorn[standard]==0.34.0

# HTTP Client (for future API integrations)
httpx==0.28.1

# Local LLM
llama-cpp-python>=0.2.23
huggingface_hub>=0.19.0

# Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0

# Document Processing
pypdf>=3.17.0

# Text Analysis
rank_bm25>=0.2.2
scikit-learn>=1.3.0

# Testing Dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Phase 2.1: Database Infrastructure
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0

# Phase 2.2: Enhanced NLP & Semantic Analysis
spacy==3.7.2
sentence-transformers==2.2.2
nltk==3.9
scikit-learn==1.5.0

# Phase 3.1: Anthropic Claude API Integration
anthropic>=0.40.0

# Phase 4: Authentication & Authorization
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# Phase 5: Production Features (Export & Templates)
reportlab==4.0.7
python-docx==1.1.0
Pillow==10.3.0
# jinja2 already included with FastAPI

# Future Phase Dependencies (not installed yet):
# Phase 5+: matplotlib>=3.8.2, pandas>=2.1.4

# Python Version: 3.10+
# Install with: pip install -r requirements.txt
reportlab
python-docx
